Add files via upload
The application is designed to promote tourism towards the cultural and natural heritage of Odisha, India. It offers complete details of all the districts of Odisha, alongside their prominent tourist places. The project also includes functionalities for train ticket bookings, hotel bookings, and travel itinerary planning, thus serving as an all-in-one system for travelers.